% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{concordance_stats}
\alias{concordance_stats}
\title{Retrieve concordance statistics}
\usage{
concordance_stats(p)
}
\arguments{
\item{p}{A plot from [ctdna_concordance()].}
}
\value{
A named list of the concordance statistics and the contingency table.
}
\description{
Returns the agreement statistics from the last [ctdna_concordance()] call
(contingency table, % agreement, kappa, weighted kappa, McNemar p,
chi-square p, Cramer's V, N) -- a friendlier alternative to reading plot
attributes.
}
\examples{
sim  <- ctdna_make_mock_study(n_patients = 30, seed = 1)
prep <- ctdna_prepare(infinity_report = sim$infinity_report,
                      tf_change = sim$tf_change, clinical = sim$clinical, verbose = FALSE)
concordance_stats(ctdna_concordance(prep))
}
