% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{longitudinal_stats}
\alias{longitudinal_stats}
\title{Retrieve longitudinal comparison tables}
\usage{
longitudinal_stats(p)
}
\arguments{
\item{p}{A plot from \code{ctdna_plot_longitudinal()}.}
}
\value{
A list with \code{$within_panel} and \code{$between_panels}.
}
\description{
Retrieve longitudinal comparison tables
}
