% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdnaTM.R
\name{print.ctdna_oncoprint}
\alias{print.ctdna_oncoprint}
\title{Print method for ctdna_oncoprint}
\usage{
\method{print}{ctdna_oncoprint}(x, ...)
}
\arguments{
\item{x}{a ctdna_oncoprint object.}

\item{...}{unused.}
}
\description{
Print method for ctdna_oncoprint
}
