% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdnaTM.R
\name{print.ctdna_pipeline_result}
\alias{print.ctdna_pipeline_result}
\title{Print method for pipeline output}
\usage{
\method{print}{ctdna_pipeline_result}(x, ...)
}
\arguments{
\item{x}{A \code{ctdna_pipeline_result} object.}

\item{...}{Ignored.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print method for pipeline output
}
