% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule-operators}
\alias{rule-operators}
\alias{&.ctdna_rule}
\alias{|.ctdna_rule}
\alias{!.ctdna_rule}
\title{Logical operators for ctdna_rule objects}
\usage{
\method{&}{ctdna_rule}(e1, e2)

\method{|}{ctdna_rule}(e1, e2)

\method{!}{ctdna_rule}(e1)
}
\arguments{
\item{e1, e2}{ctdna_rule objects.}
}
\value{
A combinator ctdna_rule object.
}
\description{
\code{`&`}, \code{`|`}, and \code{`!`} are overloaded for
\code{ctdna_rule} objects, providing fluent infix syntax equivalent
to \code{\link{allOf}}, \code{\link{anyOf}}, and \code{\link{not}}.
}
\examples{
r_path <- rule(ClinVar = c("Pathogenic","Likely_pathogenic"))
r_rare <- rule(gnomAD_AF = list(op = "<", value = 0.01))
r_path & r_rare        # equivalent to allOf(r_path, r_rare)
r_path | r_rare        # equivalent to anyOf(r_path, r_rare)
!r_path                # equivalent to not(r_path)
}
