% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ctdna_cohen_kappa}
\alias{ctdna_cohen_kappa}
\title{Cohen's kappa from a contingency table}
\usage{
ctdna_cohen_kappa(tab)
}
\arguments{
\item{tab}{Square table or matrix.}
}
\value{
Numeric kappa.
}
\description{
Cohen's kappa from a contingency table
}
