% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ctdna_concordance_test}
\alias{ctdna_concordance_test}
\title{Concordance summary for a 2-way table}
\usage{
ctdna_concordance_test(tab, method = "both")
}
\arguments{
\item{tab}{Contingency table.}

\item{method}{One of `"both","kappa","agreement","mcnemar","none"`.}
}
\value{
list with named numeric components.
}
\description{
Concordance summary for a 2-way table
}
