% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ctdna_pairwise_ranksum}
\alias{ctdna_pairwise_ranksum}
\title{All-pairs comparison test}
\usage{
ctdna_pairwise_ranksum(data, value, group, method = "wilcox")
}
\arguments{
\item{data, value, group}{See [ctdna_run_group_test()].}

\item{method}{`"wilcox"` (default) or `"t"`.}
}
\value{
Data frame: group1, group2, p, n1, n2.
}
\description{
All-pairs comparison test
}
